Add wrapper script to nbdiff Synapse notebooks
nbdime is a common tool used to diff Jupyter notebooks when using
version control tools such as git, as such diffs are otherwise quite
hard to read. Jupyter notebooks are stored as a pretty straightforward
json format. Synapse notebooks work similarly, but slight differences
require one to adapt the notebook structure before passing them to the
nbdime tools in order to generate diffs.

This wrapper script parses relevant notebooks based on provided git
references before passing the modified/new/deleted notebooks according
to the git diff to the nbdiff-web tool, in order to display the diffs
in an easy to read manner.